Started to use size_t, uintptr_t, ... consequently. Its done in kernel/src

+++ include/sys/dbg/cmd/file.h
@@ -29,6 +29,6 @@
  * @param argv the arguments
  * @return 0 on success
  */
-s32 cons_cmd_file(s32 argc,char **argv);
+int cons_cmd_file(size_t argc,char **argv);
 
 #endif /* FILE_H_ */
