+++ include/sys/mem/swapmap.h
@@ -30,34 +30,34 @@
  *
  * @param swapSize the size of the swap-device in bytes
  */
-void swmap_init(u32 swapSize);
+void swmap_init(size_t swapSize);
 
 /**
  * Allocates 1 block on the swap-device
  *
  * @return the starting block on the swap-device or INVALID_BLOCK if no free space is left
  */
-u32 swmap_alloc(void);
+uint swmap_alloc(void);
 
 /**
  * @param block the block-number
  * @return true if the given block is used
  */
-bool swmap_isUsed(u32 block);
+bool swmap_isUsed(uint block);
 
 /**
  * Determines the free space in the swapmap
  *
  * @return the free space in bytes
  */
-u32 swmap_freeSpace(void);
+size_t swmap_freeSpace(void);
 
 /**
  * Free's the given block
  *
  * @param block the block to free
  */
-void swmap_free(u32 block);
+void swmap_free(uint block);
 
 
 #if DEBUGGING
